Tried making a map
So I tried setting up the map. Added a property called vertical and
horizontal coordinates to the player class and a couple methods to
change the coordinates when they type north/south/east/west.  There's
another property called rmnum (room number) so if the player has certain
coordinates, this value changes to an integer from 0-8, each
corresponding to a different room. I then moved all the locations from
the player class into a list so that when the player character has a
certain rmnum value, their currentRoom changes to a corresponding
location in the list.
Thing is, now I get a null reference exception error on the player look
command. Even though it was working before.